api: Move to sub-package.
Split into sub-modules.

Split tests.
