+++ tests/func/test_update.py
@@ -4,7 +4,7 @@ import pytest
 
 from dvc.dvcfile import Dvcfile
 from dvc.exceptions import InvalidArgumentError
-from tests.unit.fs.test_dvc import make_subrepo
+from dvc.testing.tmp_dir import make_subrepo
 
 
 @pytest.mark.parametrize("cached", [True, False])
